Calculate discount by average score

+++ TopEducationApp/src/main/java/TopEducation/TopEducationApp/entities/StudentEntity.java
@@ -43,6 +43,6 @@ public class StudentEntity {
     private int graduationYear;
 
     // Average grade in tests
-    private double averageGrade;
+    private int averageGrade;
 
 }
